model/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
pydantic>=2.6,<3.0
httpx>=0.28,<1.0
PyNaCl>=1.5,<2.0
